backend/requirements.txt
fastapi==0.116.2
uvicorn==0.35.0
motor==3.7.1
pydantic==2.11.9
python-dotenv==1.1.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.1